+++ service-proxy-jakarta/src/test/java/no/mnemonic/services/common/api/proxy/server/ServiceProxyTest.java
@@ -10,7 +10,12 @@ import no.mnemonic.commons.utilities.collections.ListUtils;
 import no.mnemonic.services.common.api.ResultSet;
 import no.mnemonic.services.common.api.ServiceSession;
 import no.mnemonic.services.common.api.ServiceSessionFactory;
-import no.mnemonic.services.common.api.proxy.*;
+import no.mnemonic.services.common.api.proxy.ResultSetImpl;
+import no.mnemonic.services.common.api.proxy.ServiceProxyRequestContext;
+import no.mnemonic.services.common.api.proxy.TestArgument;
+import no.mnemonic.services.common.api.proxy.TestException;
+import no.mnemonic.services.common.api.proxy.TestService;
+import no.mnemonic.services.common.api.proxy.Utils;
 import no.mnemonic.services.common.api.proxy.messages.ServiceRequestMessage;
 import no.mnemonic.services.common.api.proxy.messages.ServiceResponseMessage;
 import no.mnemonic.services.common.api.proxy.serializer.Serializer;
@@ -30,14 +35,27 @@ import java.io.IOException;
 import java.net.URI;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.*;
+import java.util.Map;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.Future;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
+import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static no.mnemonic.commons.utilities.collections.ListUtils.list;
+import static no.mnemonic.commons.utilities.collections.MapUtils.map;
+import static no.mnemonic.commons.utilities.collections.MapUtils.pair;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.lenient;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 class ServiceProxyTest {
@@ -112,7 +130,7 @@ class ServiceProxyTest {
     //start 10 requests (with a tiny offset between each request, to avoid filling the threadpool immediately, which gives an unstable test
     List<Future<Response>> responses = ListUtils.list();
     for (int i = 0; i < 10; i++) {
-      responses.add(exec.submit(() -> invoke(false, "getString", "stringArg")));
+      responses.add(exec.submit(() -> invoke(false, "getString", map(), "stringArg")));
       Thread.sleep(100);
     }
 
@@ -141,10 +159,38 @@ class ServiceProxyTest {
   void testSimpleInvocation() throws IOException, TestException {
     proxy.startComponent();
     when(service.getString(any())).thenReturn("returnString");
-    Response response = invoke(false, "getString", "stringArg");
+    Response response = invoke(false, "getString", map(), "stringArg");
     assertEquals(Utils.HTTP_OK_RESPONSE, response.code);
     verify(service).getString("stringArg");
     assertEquals("returnString", readResponse(response));
+  }
+
+  @Test
+  void testReadServiceRequestContextClientIP() throws IOException, TestException {
+    AtomicReference<String> clientIP = new AtomicReference<>();
+    proxy.startComponent();
+    when(service.getString(any())).thenAnswer(i->{
+      clientIP.set(ServiceProxyRequestContext.get().getClientIP());
+      return "string";
+    });
+    Response response = invoke(false, "getString", map(), "stringArg");
+    assertEquals(Utils.HTTP_OK_RESPONSE, response.code);
+    //Verify that the clientIP is injected into the Request context
+    assertEquals("127.0.0.1", clientIP.get());
+  }
+
+  @Test
+  void testReadServiceRequestContextReadsForwardedClientIP() throws IOException, TestException {
+    AtomicReference<String> clientIP = new AtomicReference<>();
+    proxy.startComponent();
+    when(service.getString(any())).thenAnswer(i->{
+      clientIP.set(ServiceProxyRequestContext.get().getClientIP());
+      return "string";
+    });
+    //Inject a RFC7239 Forwarded header to test that the ServiceRequestContext is populated with this IP
+    Response response = invoke(false, "getString", map(pair("Forwarded", "for=10.0.1.1")), "stringArg");
+    assertEquals(Utils.HTTP_OK_RESPONSE, response.code);
+    assertEquals("10.0.1.1", clientIP.get());
   }
 
   @Test
@@ -154,7 +200,7 @@ class ServiceProxyTest {
             .setIterator(list("a", "b", "c").iterator())
             .setLimit(5).setOffset(10).setCount(20)
             .build());
-    Response response = invoke(true, "getResultSet", "stringArg");
+    Response response = invoke(true, "getResultSet", map(), "stringArg");
     assertEquals(Utils.HTTP_OK_RESPONSE, response.code);
     verify(service).getResultSet("stringArg");
     ResultSet<Object> resultSet = readResultSet(response);
@@ -168,7 +214,7 @@ class ServiceProxyTest {
   void testServiceException() throws IOException, TestException {
     proxy.startComponent();
     when(service.getString(any())).thenThrow(TestException.class);
-    Response response = invoke(false, "getString", "stringArg");
+    Response response = invoke(false, "getString", map(), "stringArg");
     assertEquals(Utils.HTTP_OK_RESPONSE, response.code);
     verify(service).getString("stringArg");
     assertInstanceOf(TestException.class, readException(response));
@@ -178,7 +224,7 @@ class ServiceProxyTest {
   void testLargeMessage() throws IOException, TestException {
     proxy.startComponent();
     String largeString = Stream.generate(() -> "a").limit(MAX_READ_STRING_LENGTH / 2).collect(Collectors.joining());
-    Response response = invoke(false, "getString", largeString);
+    Response response = invoke(false, "getString", map(), largeString);
     assertEquals(Utils.HTTP_OK_RESPONSE, response.code);
     verify(service).getString(largeString);
   }
@@ -187,7 +233,7 @@ class ServiceProxyTest {
   void testTooLargeMessage() throws IOException, TestException {
     proxy.startComponent();
     String largeString = Stream.generate(() -> "a").limit(MAX_READ_STRING_LENGTH + 10).collect(Collectors.joining());
-    Response response = invoke(false, "getString", largeString);
+    Response response = invoke(false, "getString", map(), largeString);
     assertEquals(Utils.HTTP_ERROR_RESPONSE, response.code);
     verify(service, never()).getString(any());
   }
@@ -217,19 +263,21 @@ class ServiceProxyTest {
     return serializer.deserializeB64(responseMessage.getException());
   }
 
-  private Response invoke(boolean resultset, String method, Object... arguments) throws IOException {
+  private Response invoke(boolean resultset, String method, Map<String, String> headers, Object... arguments) throws IOException {
     ServiceRequestMessage.ServiceRequestMessageBuilder requestBuilder = ServiceRequestMessage.builder()
             .setSerializerID(serializer.serializerID());
     if (arguments != null) {
       for (Object arg : arguments) {
-        requestBuilder
-                .addArgument(arg.getClass(), serialize(arg));
+        requestBuilder.addArgument(arg.getClass(), serialize(arg));
       }
     }
     try (CloseableHttpClient httpClient = HttpClientBuilder.create().build()) {
       HttpPost request = new HttpPost(URI.create(
               String.format("%s:%d/service/v1/%s/%s/%s", BASE_URL, bulkPort, TestService.class.getName(), resultset ? "resultset" : "single", method)
       ));
+      for (Map.Entry<String, String> e : headers.entrySet()) {
+        request.setHeader(e.getKey(), e.getValue());
+      }
       request.setEntity(new StringEntity(MAPPER.writeValueAsString(requestBuilder.build())));
       return httpClient.execute(request, resp -> new Response(resp.getCode(), StreamUtils.readFullStream(resp.getEntity().getContent(), true)));
     }